Move API endpoint into main.py
