+++ src/card_genius/main.py
@@ -1,6 +1,8 @@
 from fastapi import FastAPI
 from sqlmodel import Session, SQLModel, create_engine, select
 
+from models import Contact
+
 sqlite_file_name = "database.db"
 sqlite_url = f"sqlite:///{sqlite_file_name}"
 
@@ -20,3 +22,11 @@ def on_startup():
 @app.get("/")
 async def root():
     return {"message": "Hello World"}
+
+@app.post("/contacts/")
+def create_contact(contact: Contact):
+    with Session(engine) as session:
+        session.add(contact)
+        session.commit()
+        session.refresh(contact)
+        return contact
